correcció validació
